fix: remove DeviceOptionsAnyInternal type from public api
