+++ packages/timeline-state-resolver/src/__tests__/lib.ts
@@ -1,6 +1,17 @@
-import { DeviceOptionsAnyInternal } from '../conductor'
+import type { DeviceOptionsAny } from 'timeline-state-resolver-types'
 import { ConnectionManager } from '../service/ConnectionManager'
 import { MockTime } from './mockTime'
+import type { DeviceOptionsSisyfosInternal } from '../integrations/sisyfos'
+import type { DeviceOptionsVMixInternal } from '../integrations/vmix'
+import type { DeviceOptionsVizMSEInternal } from '../integrations/vizMSE'
+import type { DeviceOptionsCasparCGInternal } from '../integrations/casparCG'
+
+export type DeviceOptionsAnyInternal =
+	| DeviceOptionsAny
+	| DeviceOptionsSisyfosInternal
+	| DeviceOptionsVMixInternal
+	| DeviceOptionsVizMSEInternal
+	| DeviceOptionsCasparCGInternal
 
 /**
  * Just a wrapper to :any type, to be used in tests only
@@ -43,7 +54,7 @@ export async function addConnections(
 
 export async function removeConnections(
 	connManager: ConnectionManager,
-	connections: Record<string, DeviceOptionsAnyInternal>,
+	connections: Record<string, DeviceOptionsAny>,
 	toBeRemoved: string[]
 ): Promise<void> {
 	const addedConns: string[] = []
